# sycl: quantize and reorder the input to q8_1 when reorder is enabled