… ggml/src/ggml-sycl/vecdotq.hpp